mm/memblock.c: WARN_ON when flags differs from overlap region
Each memblock_region has flags to indicates the Node ID of this range. 
For the overlap case, memblock_add_range() inserts the lower part and
leave the upper part as indicated in the overlapped region.

If the flags of the new range differs from the overlapped region, the
information recorded is not correct.

This patch adds a WARN_ON when the flags of the new range differs from the
overlapped region.

Signed-off-by: Wei Yang <weiyang@linux.vnet.ibm.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>