ocfs2: export ocfs2_kset for online file check
When there are errors in the ocfs2 filesystem, they are usually
accompanied by the inode number which caused the error.  This inode number
would be the input to fixing the file.  One of these options could be
considered: A file in the sys filesytem which would accept inode numbers. 
This could be used to communication back what has to be fixed or is fixed.

You could write:

	$# echo "CHECK <inode>" > /sys/fs/ocfs2/devname/filecheck
or

	$# echo "FIX <inode>" > /sys/fs/ocfs2/devname/filecheck



This patch (of 5):

Export ocfs2_kset object from ocfs2_stackglue kernel module, then online
file check code will create the related sysfiles under ocfs2_kset object.

Signed-off-by: Gang He <ghe@suse.com>
Reviewed-by: Goldwyn Rodrigues <rgoldwyn@suse.de>
Cc: Mark Fasheh <mfasheh@suse.com>
Cc: Joel Becker <jlbec@evilplan.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>