page-flags: rectify forward declaration
compound_head is defined as inline in page-flags.h but in the forward
declaration of compound_head in the same file missed "inline".  As a
result we got plenty of build warnings while building for some
architecture like avr32.  The warning showed as: warning: 'compound_head'
declared inline after being called.  warning: previous declaration of
'compound_head' was here

Signed-off-by: Sudip Mukherjee <sudip@vectorindia.org>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>