Revert "mm: take i_mmap_lock in unmap_mapping_range() for DAX"
This series reverts some recent changes to the locking scheme in DAX
introduced by these two commits:

commit 843172978bb9 ("dax: fix race between simultaneous faults")
commit 46c043ede471 ("mm: take i_mmap_lock in unmap_mapping_range() for DAX")

It also temporarily disables the newly added DAX PMD fault path because of
a deadlock.  We are re-working the way that DAX does its locking for v4.4,
so for now just disable DAX PMD faults and fall back to PAGE_SIZE faults
to make sure we don't hit this deadlock.

This patch (of 3):

This reverts commits 46c043ede4711 ("mm: take i_mmap_lock in
unmap_mapping_range() for DAX") and 8346c416d17bf ("dax: fix NULL pointer
in __dax_pmd_fault()").

These commits introduced a number of deadlocks and other issues, and need
to be reverted for the v4.3 kernel.  The list of issues in DAX after these
commits (some newly introduced by the commits, some preexisting) can be
found here:

https://lkml.org/lkml/2015/9/25/602 (Subject: "Re: [PATCH] dax: fix
deadlock in __dax_fault").

This revert keeps the PMEM API changes to the zeroing code in
__dax_pmd_fault(), which were added by d77e92e270ed ("dax: update PMD
fault handler with PMEM API")

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>
Cc: Alexander Viro <viro@zeniv.linux.org.uk>
Cc: Matthew Wilcox <willy@linux.intel.com>
Cc: Dan Williams <dan.j.williams@intel.com>
Cc: Dave Chinner <david@fromorbit.com>
Cc: Jan Kara <jack@suse.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>